Move app icons definition to core-ui
