+++ core-ui/src/commonMain/kotlin/org/pointyware/timer/ui/theme/Theme.kt
@@ -1,0 +1,6 @@
+package org.pointyware.timer.ui.theme
+
+import androidx.compose.material.icons.Icons
+
+val TimerIcons = Icons.TwoTone
+val TimerIconsAutoMirrored = Icons.AutoMirrored.TwoTone
